Move functionality into ListFilter
